Move tests to correct package
