+++ src/test/java/com/github/treescrub/spedran/data/PlatformTest.java
@@ -1,10 +1,10 @@
-package com.github.treescrub.spedran;
+package com.github.treescrub.spedran.data;
 
+import com.github.treescrub.spedran.JSONLoader;
 import kong.unirest.json.JSONObject;
 import org.junit.jupiter.api.Test;
-import com.github.treescrub.spedran.data.Platform;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class PlatformTest {
