+++ src/test/java/com/github/treescrub/spedran/data/game/GameTest.java
@@ -1,7 +1,6 @@
-package com.github.treescrub.spedran.game;
+package com.github.treescrub.spedran.data.game;
 
 import com.github.treescrub.spedran.JSONLoader;
-import com.github.treescrub.spedran.data.game.Game;
 import kong.unirest.json.JSONObject;
 import org.junit.jupiter.api.Test;
