+++ src/test/java/com/github/treescrub/spedran/data/run/RunStatusTest.java
@@ -1,8 +1,6 @@
-package com.github.treescrub.spedran.run;
+package com.github.treescrub.spedran.data.run;
 
 import com.github.treescrub.spedran.JSONLoader;
-import com.github.treescrub.spedran.data.run.RunStatus;
-import com.github.treescrub.spedran.data.run.SubmissionStatus;
 import kong.unirest.json.JSONObject;
 import org.junit.jupiter.api.Test;
