+++ src/test/java/com/github/treescrub/spedran/data/variables/VariableValueTest.java
@@ -1,7 +1,6 @@
-package com.github.treescrub.spedran.variables;
+package com.github.treescrub.spedran.data.variables;
 
 import com.github.treescrub.spedran.JSONLoader;
-import com.github.treescrub.spedran.data.variables.VariableValue;
 import kong.unirest.json.JSONObject;
 import org.junit.jupiter.api.Test;
